requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.26.137
botocore==1.29.137
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.33.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-celery-results==2.5.1
django-cors-headers==4.0.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
drf-yasg==1.21.5
eventlet==0.33.3
ffmpeg-python==0.2.0
future==0.18.3
greenlet==2.0.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
MarkupSafe==2.1.2
mysqlclient==2.1.1
packaging==23.1
Pillow==9.5.0
prompt-toolkit==3.0.38
PyJWT==2.7.0
pymongo==3.12.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
redis==4.5.5
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
six==1.16.0
SQLAlchemy==2.0.15
sqlparse==0.4.4
typing-extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
